[BE] feat: 팔로우기반, 마이페이지기반 에서 태그검색

See merge request s08-ai-image-sub2/S08P22A402!182

+++ backend/src/main/java/a402/FaST/model/PK/ArticleHasTagPK.java
@@ -9,6 +9,6 @@ import java.io.Serializable;
 @NoArgsConstructor
 public class ArticleHasTagPK implements Serializable {
 
-    private int tag;
+    private Long tag;
     private int article;
 }
